+++ server/impl/src/main/java/com/walmartlabs/concord/server/org/inventory/InventoryResourceImpl.java
@@ -104,7 +104,6 @@ public class InventoryResourceImpl implements InventoryResource, Resource {
         UUID teamId = ResourceAccessUtils.getTeamId(orgDao, teamDao, org.getId(), entry);
 
         inventoryManager.updateAccessLevel(inventoryId, teamId, entry.getLevel());
-
         return new GenericOperationResult(OperationResult.UPDATED);
     }